hotfix(scheduler): @Modifying long 리턴 타입을 int 로 (E2 P1 컴플라이언스, 매일 03:00 개인정보 파기 실패 해소)
PersonalDataAccessLogRepository.deleteByTenantIdAndAccessTimeBefore
- Spring Data JPA 3.x strict 검증에서 @Modifying + long 리턴은 IllegalArgumentException 발생
- 매일 03:00 KST 1년 경과 개인정보 접근 로그 파기가 7일 연속 실패하여 누적 중
- long → int 변경, 호출 측 PersonalDataDestructionService:301 의 (int) deletedCount 캐스팅은 그대로 유지 (호환)

+++ src/main/java/com/coresolution/consultation/repository/PersonalDataAccessLogRepository.java
@@ -97,7 +97,7 @@ public interface PersonalDataAccessLogRepository extends JpaRepository<PersonalD
      */
     @Modifying(clearAutomatically = true, flushAutomatically = true)
     @Query("DELETE FROM PersonalDataAccessLog p WHERE p.tenantId = :tenantId AND p.accessTime < :cutoffDate")
-    long deleteByTenantIdAndAccessTimeBefore(@Param("tenantId") String tenantId, @Param("cutoffDate") LocalDateTime cutoffDate);
+    int deleteByTenantIdAndAccessTimeBefore(@Param("tenantId") String tenantId, @Param("cutoffDate") LocalDateTime cutoffDate);
     
     /**
      * 테넌트별 특정 개인정보 유형의 접근 로그 조회 (tenantId 필터링)
